fix(business): corriger 5 bugs UI détectés en revue de pages
- Sidebar: analytics tabs jamais actifs (pathname ne contient pas les query params) → useSearchParams pour comparer tabActif
- LicencesClient: parseInt → parseFloat pour conserver les décimales du prix
- CommandesClient: sousTotal et remise sans .toFixed(2) dans la PreviewModal
- AbonnementsClient: signe € avant le montant → standardiser après (X.XX€)
- TabVentes: carte "CA Net" en doublon (déjà présente en sous-titre de CA Brut)

+++ app/dashboard/business/licences/_components/LicencesClient.tsx
@@ -68,7 +68,7 @@ export default function LicencesClient({ licences: initial }: { licences: Licenc
       setLicences(l => l.map(x => x.id === licence.id ? {
         ...x,
         nom: form.nom,
-        prix: parseInt(form.prix),
+        prix: parseFloat(form.prix),
         streams_limite: form.streams_limite ? parseInt(form.streams_limite) : null,
       } : x))
       setEditId(null)
